[model, utils] Separate concurrent queue and test.
