+++ cpp_tests/test-main.cpp
@@ -1,0 +1,5 @@
+#define BOOST_TEST_MODULE Test Accelerator
+#include <boost/test/included/unit_test.hpp>
+
+#include "test-accelerator.h"
+#include "test-concurrent-queue.h"
